NgoVanQuang_20183813: sua get/set trong usermodel va rolemodel

+++ Final/Backend/QuanLyKTX/src/main/java/web/nhom8/quanlyktx/model/RoleModel.java
@@ -21,7 +21,7 @@ public class RoleModel {
         super();
     }
 
-    public int getRoleId() {
+    public long getRoleId() {
         return RoleId;
     }
